Back multiusuario funcionando, rol arreglado, claves foráneas correctas, get por portafolio funcionando, creacion de usuario funcionando, login funcionando, potenciar esta app!

+++ src/main/java/com/ProyectoFinalArgProg/crud/repository/EducacionRepository.java
@@ -20,7 +20,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EducacionRepository extends JpaRepository<Educacion, Long> {
+public interface EducacionRepository extends JpaRepository<Educacion, Integer> {
 
     @Query(value="SELECT * FROM educacion WHERE id_usuario =:idUsuario" , nativeQuery = true)
     Set<Educacion> buscarTodaLaEducacionPorIdUsuario(@Param("idUsuario") Integer idUsuario);
